+++ packages/@expo/cli/src/start/server/metro/__tests__/withMetroMultiPlatform.test.ts
@@ -4,12 +4,25 @@ import type { ConfigT } from '@expo/metro/metro-config';
 import type { CustomResolutionContext, Resolution } from '@expo/metro/metro-resolver';
 import { vol } from 'memfs';
 import assert from 'node:assert';
+import resolveFrom from 'resolve-from';
 
 import { AutolinkingModuleResolverInput } from '../createExpoAutolinkingResolver';
 import { shouldCreateVirtualCanary, shouldCreateVirtualShim } from '../externals';
 import { getNodejsExtensions, withExtendedResolver } from '../withMetroMultiPlatform';
 
-const asMetroConfig = (config: Partial<ConfigT> = {}): ConfigT => config as any;
+jest.mock('resolve-from', () => {
+  const actual = jest.requireActual<typeof import('resolve-from')>('resolve-from');
+  const resolve = jest.fn(actual) as any as typeof actual;
+  resolve.silent = jest.fn(actual.silent);
+  return resolve;
+});
+
+const asMetroConfig = (config: Partial<ConfigT> = {}): ConfigT => ({
+  ...(config as any),
+  transformer: {
+    asyncRequireModulePath: 'expo/internal/async-require-module',
+  },
+});
 
 class FailedToResolveNameError extends Error {
   extraPaths: string[] = [];
@@ -97,6 +110,10 @@ function getResolveFunc() {
   const metroResolver: typeof import('@expo/metro/metro-resolver') = require('@expo/metro/metro-resolver');
   return metroResolver.resolve;
 }
+
+beforeEach(() => {
+  jest.restoreAllMocks();
+});
 
 describe(withExtendedResolver, () => {
   function mockMinFs() {
@@ -342,13 +359,9 @@ describe(withExtendedResolver, () => {
   });
 
   it(`resolves to @expo/vector-icons on any platform`, async () => {
-    vol.fromJSON(
-      {
-        'node_modules/@react-native/assets-registry/registry.js': '',
-        'node_modules/@expo/vector-icons/index.js': '',
-      },
-      '/root/'
-    );
+    jest.mocked(resolveFrom.silent).mockImplementation((_from, moduleId) => {
+      return moduleId === '@expo/vector-icons' ? 'node_modules/@expo/vector-icons' : undefined;
+    });
 
     ['ios', 'web'].forEach((platform) => {
       const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/root/' }), {
@@ -371,13 +384,9 @@ describe(withExtendedResolver, () => {
   });
 
   it(`resolves nested imports to @expo/vector-icons on any platform`, async () => {
-    vol.fromJSON(
-      {
-        'node_modules/@react-native/assets-registry/registry.js': '',
-        'node_modules/@expo/vector-icons/index.js': '',
-      },
-      '/root/'
-    );
+    jest.mocked(resolveFrom.silent).mockImplementation((_from, moduleId) => {
+      return moduleId === '@expo/vector-icons' ? 'node_modules/@expo/vector-icons' : undefined;
+    });
 
     ['ios', 'web'].forEach((platform) => {
       const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/root/' }), {
@@ -400,12 +409,7 @@ describe(withExtendedResolver, () => {
   });
 
   it(`does not alias react-native-vector-icons if @expo/vector-icons is not installed`, async () => {
-    vol.fromJSON(
-      {
-        'node_modules/@react-native/assets-registry/registry.js': '',
-      },
-      '/root/'
-    );
+    jest.mocked(resolveFrom.silent).mockReturnValue(undefined);
 
     ['ios', 'web'].forEach((platform) => {
       const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/root/' }), {
@@ -732,6 +736,67 @@ describe(withExtendedResolver, () => {
       '/node_modules/react-native-web/dist/cjs/exports/AppRegistry/AppContainer.js',
       'web'
     );
+  });
+
+  it('aliases assets registry to virtual shim', async () => {
+    vol.fromJSON(
+      {
+        'node_modules/@react-native/assets-registry/registry.js': '',
+        mock: '',
+      },
+      '/'
+    );
+
+    const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/root/' }), {
+      tsconfig: {},
+      getMetroBundler: getMetroBundlerGetter(),
+    });
+
+    const result = modified.resolver.resolveRequest!(
+      getDefaultRequestContext(),
+      '@react-native/assets-registry/registry',
+      'ios'
+    );
+
+    expect(result).toEqual({
+      filePath: '\0polyfill:assets-registry',
+      type: 'sourceFile',
+    });
+  });
+
+  it('aliases async require module to resolved path', async () => {
+    // Mock path we're expecting `asyncRequireModulePath` requests to have been replaced with
+    const expectedPath = 'node_modules/expo/internal/async-require-module.js';
+
+    vol.fromJSON(
+      {
+        'node_modules/@react-native/assets-registry/registry.js': '',
+        mock: '',
+      },
+      '/'
+    );
+
+    jest.mocked(resolveFrom.silent).mockImplementation((_from, moduleId) => {
+      return moduleId === config.transformer.asyncRequireModulePath ? expectedPath : undefined;
+    });
+
+    const config = asMetroConfig({ projectRoot: '/root/' });
+    const modified = withExtendedResolver(config, {
+      tsconfig: {},
+      getMetroBundler: getMetroBundlerGetter(),
+    });
+
+    // Requesting `asyncRequireModulePath` will replace the path with a Node-resolved path
+    const result = modified.resolver.resolveRequest!(
+      getDefaultRequestContext(),
+      config.transformer.asyncRequireModulePath,
+      'ios'
+    );
+
+    expect(result).toEqual({
+      filePath: expectedPath,
+      type: 'sourceFile',
+    });
   });
 
   describe('built-in externals', () => {